Integrate transport service

+++ apps/transport/transport-common/src/main/java/transport/common/CancelReservationEvent.java
@@ -8,6 +8,6 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 @NoArgsConstructor
 public class CancelReservationEvent extends BaseEvent {
-  private int numberOfPeople;
+  private Long numberOfPeople;
   private String userName;
 }
